added ordering system

+++ database/migrations/2020_11_12_135829_create_order_product_table.php
@@ -15,8 +15,9 @@ class CreateOrderProductTable extends Migration
     {
         Schema::create('order_product', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('order_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('order_id');
+            $table->unsignedBigInteger('product_id');
+
 
             $table
                 ->foreign('product_id')
